refactor: rename utils & types packages
