+++ packages/@wdio-electron_types/tsconfig.json
