+++ packages/@wdio-electron_utils/src/cjs/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@repo/utils",
+  "name": "@wdio-electron/utils",
   "type": "commonjs",
   "private": true
 }
