+++ packages/@wdio-electron_utils/src/constants.ts
