+++ packages/@wdio-electron_utils/src/log.ts
