+++ packages/wdio-electron-service/src/cjs/session.ts
@@ -3,7 +3,7 @@ import type { Options } from '@wdio/types';
 
 import { CJSElectronLauncher, CJSElectronService } from './classes.js';
 import { CUSTOM_CAPABILITY_NAME } from './constants.js';
-import type { ElectronServiceOptions } from '@repo/types';
+import type { ElectronServiceOptions } from '@wdio-electron/types';
 export async function init(opts: ElectronServiceOptions) {
   // CJS variants of the Launcher and Service classes are needed here
   // - which is why we are not simply doing a dynamic import of `../session.js`
