+++ packages/wdio-electron-service/src/commands/mock.ts
@@ -1,6 +1,6 @@
 import { createMock } from '../mock.js';
 import mockStore from '../mockStore.js';
-import type { ElectronMock } from '@repo/types';
+import type { ElectronMock } from '@wdio-electron/types';
 
 export async function mock(apiName: string, funcName: string): Promise<ElectronMock> {
   try {
