+++ packages/wdio-electron-service/src/index.ts
@@ -1,6 +1,6 @@
 import { browser as wdioBrowser } from '@wdio/globals';
 
-import type { ElectronServiceOptions } from '@repo/types';
+import type { ElectronServiceOptions } from '@wdio-electron/types';
 
 import { init as initSession } from './session.js';
 import ElectronLaunchService from './launcher.js';
@@ -11,4 +11,4 @@ export default ElectronWorkerService;
 
 export const browser: WebdriverIO.Browser = wdioBrowser;
 export const startElectron: (opts: ElectronServiceOptions) => Promise<WebdriverIO.Browser> = initSession;
-export * from '@repo/types';
+export * from '@wdio-electron/types';
