+++ packages/wdio-electron-service/src/session.ts
@@ -4,8 +4,8 @@ import type { Options } from '@wdio/types';
 import ElectronWorkerService from './service.js';
 import ElectronLaunchService from './launcher.js';
 import { CUSTOM_CAPABILITY_NAME } from './constants.js';
-import log from '@repo/utils/log';
-import type { ElectronServiceOptions } from '@repo/types';
+import log from '@wdio-electron/utils/log';
+import type { ElectronServiceOptions } from '@wdio-electron/types';
 
 export async function init(opts: ElectronServiceOptions) {
   const testRunnerOpts = opts as Options.Testrunner;
