+++ packages/wdio-electron-service/test/mock.spec.ts
@@ -3,7 +3,7 @@ import { isAsyncFunction } from 'node:util/types';
 import { describe, it, expect, beforeEach, vi, Mock } from 'vitest';
 
 import { createMock } from '../src/mock.js';
-import { ElectronInterface, ElectronType } from '@repo/types';
+import { ElectronInterface, ElectronType } from '@wdio-electron/types';
 
 let mockFn: Mock;
